🐛 soften requirement on policy spec keys (#60)
So far we were checking for the MRN to be a a policy MRN during the
validation.

However, there are cases where we may associate dynamic policies with
assets, whose MRNs are not policy MRNs. Only check for MRN in these
places of validation and. Note: When it comes to printing, however, we
do look for specific policies, so the `IsPolicyMrn` method is still
necessary to filter results from the report and not print things that
are not policies.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>