⭐ restore scan -o report
Note: After the migration we are loosing some context from policy bundles, so we don't always have the asset policy available in the bundle. There is a new shared call in the resolved policy, which can help find the top policies given any bundle. This also means that we can hand it any random bundle we are about and given a resolved policy find the root policies that were pulled from the bundle. It helps render policies in order i.e. embedded policies are can be tackled further down. We only need this approach when looking at policies in their hierarchy, as for everything else we can just take the bundle and the results. Since the report displays policies in their hierarchy, it required these new calls.
Signed-off-by: Dominik Richter <dominik.richter@gmail.com>