📃 improve readme structure
Great feedback from @IronMania: We are mentioning custom policies very early in the process, where it's more helpful to talk about scaling and the shell, before diving into custom policies. Restructured the readme and moved the custom policy part to the back. We are going to create a shorter develop-centric policy-writing guide as well, which will serve as a more lightweight intro.

Additionally move  out of the  subfolder, so they are found more easily by just looking at the repo.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>