🌠🛑 policy v2 (#301)
WIP: the proto files need rebuilding after the cnquery PRs are rebased
(see https://github.com/mondoohq/cnquery/pull/813 for the chain of PRs)

This change builds on top of the [query
v2](https://github.com/mondoohq/cnquery/pull/713) and
[properties](https://github.com/mondoohq/cnquery/pull/724) changes in
cnquery. It updates the structure for policies to be more flexible and
standalone moving forward.

It also introduces `impact` which will start deprecating `severity`
moving forward. This is purely semantics, none of the underlying logic
has changed. The only reason we are making this change now has been a
ton of feedback from friends in the community, who pointed out that
severity occupies a somewhat different meaning (that is very very tied
to CVSS).

One of the biggest changes in this PR is the fact that we are
deduplicating a lot of structures, most notable that of the `Mquery`
(i.e. query metadata and MQL code). The core implementation now resides
in cnquery and cnspec simply builds on top of it. A few others
structures are affected as well, but should be largely unnoticeable.

We are also overhauling the property system and making it fully usable
on the CLI. You can now specify properties via e.g.:

```
cnspec scan -f my.mql.yaml --props "homeFolder='/home/user'"
```



Signed-off-by: Dominik Richter <dominik.richter@gmail.com>

---------

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>
Co-authored-by: Christoph Hartmann <chris@lollyrock.com>