🐛 re-activate and fix checksum generation on bundles (#323)
... using the updated cnquery methods to merge base queries and
correctly re-generate their checksums.

The current approach had errors when it comes to base and embedded
queries. That is because previously we would only deal with what we now
call base queries and didn't have to worry too much about any overrides
that were happening, because those were limited to scoring specs.
However, the latter have now been replaced and merging queries has
become the norm. In this world, we need correct checksums for all merged
actions.

Additionally, we don't want to alter the structure of query overrides
too much, i.e. we don't want to just pull the base query into the
override and overwrite the policy group, since that would lose all
context of a dependency existing. We will revisit some of this behavior
when we get to version history of snapshotting in cnspec/cnquery, which
is coming up soon.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>