🐛 setting properties updates the checksums (#360)
The SetProps call didn't update checksums. There was also no way to
delete a property.

Since the call is only used when property values (i.e. the MQL) is set,
we can remove properties if they are called with the set call but no MQL
is provided. This call doesn't make sense when not used for overriding
values, since all other policy- or query-based properties are embedded
and don't use this call.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>