🐛 FillV7 on bundle should respect properties (#422) (#438)
also moves the deprecated example for v7 out of the `examples/` folder
into the folder where the convertor is. Added tests as well with
consistent naming, so we can delete both later on.

This PR originally had to be pulled because it broke existing v7 bundles
that did not convert correctly. Upon closer inspection, it became
visible that the mechanism had exposed even deeper flaws in the way that
properties were handled in v7. The original PR ended up submitting a
bundle which crashed at the following position:

```go
propsQueries := map[string]*Mquery{}
for k := range policyObj.Props {
	q, ok := bundle.Props[k]
	if !ok {
		return nil, nil, status.Error(codes.InvalidArgument, "policy "+policyObj.Mrn+" is referencing unknown property "+k)
	}
	propsQueries[k] = q
}
```

This means we were associating properties in the policy which were not
found in the bundle. However, before the original PR these properties
were entirely missing from policies (this is an example running
cnspec@17e2315).

-----

Here is a full example:


![image](https://user-images.githubusercontent.com/1307529/224904988-265ea9dd-f21a-4519-a0a9-b07be8d7b595.png)


![image](https://user-images.githubusercontent.com/1307529/224905021-1f2d94b3-8436-45aa-98da-f92552d25236.png)


![image](https://user-images.githubusercontent.com/1307529/224905053-7cf13c8d-1e24-47bd-8359-2eca36160ccd.png)


![image](https://user-images.githubusercontent.com/1307529/224905108-293e5ed5-1621-4fdf-8b60-0a753d4ee71a.png)

This trace is showing the `sshd-policy` that has `Props` set to `nil`
(which in v7 terms means there are no properties associated with any of
the queries, while we clearly see that one of the checks `sshd-sshd-01`
has a property associated with it in the `Query` field.

However, this policy ran fine in v7 clients, likely mistakenly so.

-----

The same with this PR has properties in the bundle and in the policy:


![image](https://user-images.githubusercontent.com/1307529/224905614-1decb927-b310-4ef0-bd6a-19d9cc1d91c8.png)


![image](https://user-images.githubusercontent.com/1307529/224905736-6d599323-fac9-4e99-9a8e-96173a42a832.png)

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>