postV8: 🧹 bundle compiler cleanup + complex example (#425)
Note: This requires https://github.com/mondoohq/cnquery/pull/1046 and an
update to go.mod

1. The bundle compiler has been updated and simplified similar to
cnquery. However, there is more work to be done here, this is only the
beginning.

2. The multi example was outdated and was updated to a new complex
example. It already includes variants, despite the fact that those are
not yet executed. The policy runs, but it will respect variants in an
upcoming fix.

---------

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>
Signed-off-by: Joel Diaz <joel@mondoo.com>
Co-authored-by: Joel Diaz <joel@mondoo.com>