🐛⭐ fix variants execution (#451)
Allows for variants in checks and queries to be properly executed in v8
(as planned).

The complex policy has been expanded with examples of variants in both
checks and queries. We ... could reduce one of those, it's getting full
in there, but it does get the point across. See
`examples/complex.mql.yaml`

You'll get this output on a Linux system with the above policy:


![image](https://user-images.githubusercontent.com/1307529/225550401-a9353e82-037d-42a9-8d88-95597ff44c15.png)

As you can see: only one of the queries survived and returned the `uname
-a` output, while the windows query variant was not run. We don't get
more info on the default CLI output and frankly it's great to not
confuse it.

Additionally only "A check only run on Linux/macOS" is executed, while
the windows variant of the check is not run. We decided to keep the
check's metadata, but should probably aggregate it better in a follow-up
(e.g. if the base variant doesn't have a title, pull it from the
parent...).

Finally, if used with json output:


![image](https://user-images.githubusercontent.com/1307529/225551095-259c5eb1-4bbc-4b76-8339-ce80d66b7c26.png)

You can see both the variant check and its parent check being returned
in the JSON data structure. Data queries are of course only returned for
the executed query itself (and in the future these will be even stronger
tied to the actual datafield and less to the query).

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>