🧹 remove RefreshLocalAssetFilters + update to cnquery methods
Local asset filters was an optimized call pre-v8 when we didn't have filters on individual queries. Now we have local filters and even filters in variant queries, which makes things a lot harder. To that end, we remove any aggregation of local filters and instead go into compute filters, which does a deep-dive.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>