Modify ComputeAssetFilters to take variants into account (#452)
The following example was not working:
```
policies:
  # This is a second policy in the same bundle
  - uid: example2
    name: Another policy
    version: "1.0.0"
    groups:
      # Additionally it defines some queries of its own
      - type: chapter
        title: Hello World
        filters: asset.platform == "arch" || asset.platform == "asdf"
        checks:
          - uid: helloworld
            variants:
              - uid: helloworld1
              - uid: helloworld2
queries:
- uid: helloworld1
  title: Hello World Query
  filters: asset.platform == "arch"
  query: |
    "hello" == "hello"
- uid: helloworld2
  filters: asset.platform == "asdf"
  title: Hello World Query
  query: |
    "hello" == "world"
```

My platform is arch. I would expect helloworld1 to run. Instead, no
queries run and I get a `U`

I'm not sure of the following:
- ComputeAssetFilters didn't look like it was looking at the queries and
checks
- Why do we need both RefreshLocalAssetFilters and ComputeAssetFilters
- Why was ComputeBundle not using ComputeAssetFilters to recalculate the
  filters

I've also changed the inmemory database implementation to make a copy of
the data it returns. I was running into issues where functions would get
data from the db and scribble on it, affecting the data in the db. Now
that I think about it, it probably should be copied on the way in as
well

requires https://github.com/mondoohq/cnquery/pull/1077

---------

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>
Co-authored-by: Dominik Richter <dominik.richter@gmail.com>