⭐️ configurable report sharing (#543)
Users can now simply share a report via:

```
cnspec scan --share-report
```

If that is not set then the user is asked:

```
Do you want to view or share this report results via Mondoo's reporting service in a browser? [Y/n]
```

To disable the the sharing and not being asked, users either

- use the `--share-report=false` flag
- set the config with the following settings

```
share_report: false # only applicable if used unauthenticated 
```

When a user runs `cnspec scan --share-report` and has disabled the
reporting in the config, the cli flag has precedence and therefore the
report will be shared.