🧹 improve linter for variants (#546)
fixes #526

With the new v8 and variants we support queries to have their own asset
filter. Therefore policy groups do not necessarily need an asset filter
anymore. We still need to see how this shapes out long term but it
definitely makes sense not to show the warning when variants are uses
since they need to have an asset filter.