🧹 allow specifying custom http timeout for serve-api cmd (#550)
The default HTTP timeout is 30s but sometimes that's not enough. For
example, if the mondoo-operator needs to clean thousands of assets, it
might take more than 30s. The operator itself sets the timeout for its
client to 50minutes or so but the scan API terminates the request on
30s.

Added a flag to allow extending this timeout when needed

Signed-off-by: Ivan Milchev <ivan@mondoo.com>