🌠 compliance frameworks (#584)
This is the first step in introducing Compliance Frameworks.

**Why Compliance?** Because the vast majority of security testing and
reporting is used for either one of two purposes: hardening your tech to
protect against attackers AND to adhere to mostly regulatory compliance.

Security is served well with policies and prioritized results. Policies,
checks, and additional data queries provide everything we need.

Compliance, however, requires a lot of structured evidence-gathering and
result aggregation across a lot of technologies and policies.
Additionally, it puts a much stronger emphasis on exception-handling.

**Why compliance frameworks?** An alternate route to this approach could
have been to just tag all security checks with identifiers from common
compliance frameworks. This is commonly good enough... until something
changes. Versioning, in this model, is tedious at best and can lead to
outright incorrect results at its worst.

Together with a lack of versioning, we have seen toolsprawl and
aggregation cause annoying headaches in the relationship of security and
compliance.

Today we are recognizing Compliance Frameworks as a codified artifact,
which derives its information from security policies, its checks and
queries, as well as other controls. As such, it can be versioned,
scaled, and extended.

The model that is introduced today allows users to extend frameworks and
create bindings between new policies and existing frameworks. Unlike
policies, which are fully maintained by their authors, compliance
frameworks are an openly extensible object, which allows other users to
create new bindings.

**Please note: This capability is experimental** until the release of v9
in the upcoming weeks. We appreciate your help in shaping this out
tremendously! Feedback very welcome!


![image](https://github.com/mondoohq/cnspec/assets/1307529/f36a315f-7ed4-4546-9cc7-022cab7e5978)

Follow-up:
- add testing
- re-introduce policy summary in compact CLI output, so we can also
print the framework in the summary

---------

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>
Co-authored-by: Christoph Hartmann <chris@lollyrock.com>