🐛 Connect compliance frameworks to the root of the graph
Compliance frameworks were not connected to the root of the graph. On
the backend, we iterate from the root to figure out all reporting jobs
that get saved. Without this, we don't save the compliance frameworks or
controls

Not done here, but we should also be preserving the framework
dependencies in the graph so that its consistent to how policies work