Preserve the framework dependencies in the reporting structure
Create reporting jobs for the chain of frameworks. For example, asset
frameworks point to the space framework which point to the global framework
which point to some actual frameworks.

One issue I ran into was that we have space/asset frameworks and
space/asset policies. I chose to not create separate reporting jobs for
those because having multiple reporting jobs with the same query is
likely to break something. Instead, the policies would just get attached
to the existing policy jobs, but the impact is set to unscored