:broom:  Incorporate framework maps into the framework checksum  (#639)
With this, frameworks graph checksums will change if any of their
framework maps change. This allows us to correctly pick up changes to
the map during building of the resolved policy. Without this, we'd just
keep picking up the cached resolved policy when the framework maps
change because the graph checksum wouldn't change