🧹 Report checks by mrn when used by controls (#672)
For frameworks reporting, we want the ability to display how checks
scored for a control. We aggregate the information by check. In some
cases, we have checks that have different MRNs but the same code. This
makes it difficult to separate these. This change modifies the reporting
structure for controls. We don't directly send the check score to the
control. Instead, a new reporting job is inserted that reports to the
query mrn. And then that reports to the control.


![rp](https://github.com/mondoohq/cnspec/assets/27443/8ceb5dc6-496d-4bd2-908e-a14f846b1184)