🐛 Fix check control resolver bug (#687)
When we have controls that map to check, the resolver creates a
reporting job from the query by its code id to query by its mrn. And
then maps that to the control. However, there was a bug in this
translation where we would pick a random check mrn from the bundle that
had that code id. Since its possible for multiple checks to have the
same id in a bundle, and not all of those checks are part of an active
policy, this would end up picking a mrn from a policy that wasn't
active.