🐛 datapoints are conditional on providers
The os provider may not exist on the run or on the system. In these cases, the method to determine the datapoint for vulnerability report or kernel info will fail. To better handle these errors, we abandon the previous approach of `MustGetOneDatapoint` via `MustCompile` and instead use the provider info to report back to users.

There may still be other reasons why the compile fails, but we will keep this focused on the primary reason in a tested environment with a working provider runtime that indicates why this fails: no provider for this code.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>