🐛 support --record on scan command
Note: there is a bug where it does not associate the correct asset, because we reconnect and the ID of the asset is wrong. Will be fixed...

Additionally, there is a bug where running this command again fails to correctly load the recording.

Finally: removed some old commented code, usually would have been in a separate PR but this will not be contentious
Signed-off-by: Dominik Richter <dominik.richter@gmail.com>