⭐ wire data queries to controls
This creates the internal wiring to collect data queries for controls.
It connects each datapoint that a query produces to all the reporting
job of the control that it is mapped to.

Follow-up: I noticed that we don't do a decent job on printing just yet,
e.g. we only print if a control passes or fails bound to its score.
That's why there isn't a good place to add datapoints yet, considering
we already print all data that is executed for a policy. Definitely
looking to improve the CLI experience.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>