⭐️ Add reporting jobs for data queries.
Until now we turned data queries's reporting jobs into datapoints and attached those to the parent RJs. With this change we instead attach the datapoints to the data queries' RJs and we do not delete them.
This means that we now have scores coming in for data queries, even though those will always be of the unscored type.

Signed-off-by: Preslav <preslav@mondoo.com>