🐛 Fix issue with policies matching against incorrect platforms (#1036)
The issues here is the way policies are selected to be included.
It checks this field called ComputedFilters of the policy.
I'm not 100% sure what this is, but from the comment, it sounds
like this has any filter that this policy uses. However, I don't
think that is useful to answer the question of whether the policy
should be included. For example, I run arch linux. If I include
the alma linux 8 cis benchmark, that policy is included. That is
because the following query is defined:

```
  - uid: cis-almalinux-8--4.2.11
    title: Ensure sshd KexAlgorithms is configured
    impact: 100
    filters: |
      file("/etc/crypto-policies/config").exists == false
      command('update-crypto-policies --show').stdout.trim == ""
...
```

This change fixes the problem by pruning out any reporting job
nodes that are policies that have no children or datapoints