🐛 extend risk factor checksum calculation
This now includes all fields and separates the checksum between the execution and content part. The RiskFactor itself only tracks the content checksum, while we use the execution checksum when we calculate the policy checksum.

Particularly this last part was missing from risk calculations and cause the policy content/execution checksums to not get updated.

I also cleaned up a weird for-loop in the referenced policies checksum calculation.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>