🐛 fix RiskFactor Checksum with Magnitude field
We ignored value changes in this field because the float was mapped (and
rounded) to an integer before adding it to the checksum. Despite having
a test for this case, the test didn't catch this case because other
changes to the RiskFactor caused changes in the checksum.

We now have a standalone method for generating data for RiskFactor tests
and additional validations that the generation produces unchanged data.
(This approach is also more efficient than a deep-clone via proto)

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>