🧹 only submit scores once we are done (#1222)
We will wait to send scores to the server until all of them have been
collected. This helps us to update all the scores with the available
risks in the client before sending them.

Follow-up: We may want to batch up the number of scores we send and
possibly split out the data from the remaining score upload if there is
a lot to send. Ultimately, the last batch (which sets the scan to done)
must contain all policies, which makes it easy to consume them in just
one request.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>