critical: fix banded scoring on variation within bands (#1385)
Variations within bands were inverted and led to incorrect results on
the edges of those variations i.e. if all scores within a band were
failing or only one was failing. The tests reflect have been added to
test out this behavior of variation within bands and will be further
expanded in the future.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>