🐛 Don't tie incorrect data to controls (#1400)
We have a case where an incorrect data query is being tied to a control
by mrn. This happens when there is duplicate code ids for different
mrns.

This is causing a resolved policy to be created where the child jobs are
not consistent with notifies. When executed, things get scored as a U.