⭐️ new `cnspec integrate` experimental command
This is an **EXPERIMENTAL** command.

Our first iteration of an automated way to create integrations in
the Mondoo platform. ✨

The first integration we are enabling is Azure, to do so, the user
can run:
```
cnspec integrate azure --space <the_space_to_create_the_integration>
```

The flow to create an integration is:

* Verify we have access to the Mondoo platform
* Verify we can access Azure
* Detect subscriptions to integrate
* More validation like, integration name, etc.
* Generation of automation code
* Write code to disk
* Run automation code with confirmation from user

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>