⭐️ integrate preflight check: role assignments
This change adds a minimalistic preflight check where we verify that the
user has the right role assignments to onboard an Azure environment.

We need one of "Privileged Role Administrator" or "Global Administrator"
=> https://learn.microsoft.com/en-us/entra/identity/role-based-access-control/permissions-reference#privileged-role-administrator

Note that we run Azure CLI commands directly so that `cnspec` doesn't
have to import the entire Azure SDK which will make the binary much
bigger, which today is unnecessary for the amount of checks we are
running.

Signed-off-by: Salim Afiune Maya <afiune@mondoo.com>