🧹 Make resolved policy execution use reporting queries (#1496)
The execution was previously implicitly tying together reporting queries
and reporting jobs. This now uses the ReportingQueries section of the
CollectorJob to do this.

Also, part of this change, I've introduced an additional reporting job
types. One represents the executed query(reporting job by code id). The
other represents a thing that is both a check and a query. This was
helpful in fixing an issue where we don't properly forward a score and
intead rescore it. This would turn skipped scores into unscored.

This last point is one of those non-backwards compatible changes. Once
the new resolved policy is used everywhere, this bug will be seen in old
clients. I don't see a way to fix it in a backwards compatible way, and
its a small enough issue that I think its fine.