Improve logic in sysctl remediations + improve auditd remediation (#1761)
* Improve logic in sysctl remediations

I noticed a lot of our steps didn't work for my Debian system because
the config was already set in the /etc/sysctl.conf file. If it's there
let's make sure to edit it. If not then we can create the .d file

Signed-off-by: Tim Smith <tsmith84@gmail.com>

* Fix some incorrect reloading

Turns out the main config is not reloaded with a reload. You have to do
a full restart

Signed-off-by: Tim Smith <tsmith84@gmail.com>

* Improve the max log file size bash script

Signed-off-by: Tim Smith <tsmith84@gmail.com>

* Add bash remediation for auditd max_log_file_action

Signed-off-by: Tim Smith <tsmith84@gmail.com>

* Add more auditd remediations

Signed-off-by: Tim Smith <tsmith84@gmail.com>

* Improve bash remediation steps

Signed-off-by: Tim Smith <tsmith84@gmail.com>

* Add steps to manual remediations

Signed-off-by: Tim Smith <tsmith84@gmail.com>

* Don't use sudo in remediations

Signed-off-by: Tim Smith <tsmith84@gmail.com>

---------

Signed-off-by: Tim Smith <tsmith84@gmail.com>