⭐ Add a new mechanism to send results to the server
This PR introduces a new way to send results to the server. Instead of
sending data through the API, we instead upload a file to a link
provided by the server.

Two new APIs are introduced: GetUploadURLs and ConfirmUploads.
- *GetUploadURLs* requests a place where it can upload the scan results
- *ConfirmUploads* is called once all files have been uploaded

This iteration uses a single file bundle. All results are written to an
SQLite database. Tables are created for scores, data, risk factors and
resources. SQLite was chosen for two reasons:
- It avoids needing to create a new file format
- SQLite deals with memory management for us. Related to the file
  format, but we for free get the ability to not have to deserialize all
  data into memory to process it. This will also be helpful during
  writes as we will not need to hold all data in memory (the graph
  executor and resource recorder would still need to be modified to
  not keep a copy of this and instead write directly to the store)

The implementation is less than ideal in some places. To avoid creating
a whole new datalake, we give the datalake a writer for results. This
lets us swap out the part that directly writes in memory with one that
can write to any store of our choosing. Things like policies, resolved
polciy, etc still get cached in the "in-memory" part of the data lake.