🐛 Convert implicit props into explicit props
cnquery v12 introduced the ability to have implicit properties.

The following failed to compile:
```
policies:
  - uid: implicit-props
    name: implicit-props policy 1
    version: "1.0.0"
    authors:
      - name: Mondoo
        email: hello@mondoo.com
    groups:
      - filters: return true
        queries:
          - uid: query-1
    props:
      - uid: xtesting
        mql: return true

queries:
  - uid: query-1
    title: testing
    mql: props.xtesting == true
```

This change converts implicit props to explicit props when the bundle is
uploaded. A query can lookup from any of its parents. It goes up the
chain in order, using the first matching prop. If one is found, we
create the prop on the query. The matching prop on the parent references
it through `for`.

The bundle above becomes:
```
policies:
  - uid: implicit-props
    name: implicit-props policy 1
    version: "1.0.0"
    authors:
      - name: Mondoo
        email: hello@mondoo.com
    groups:
      - filters: return true
        queries:
          - uid: query-1
    props:
      - mrn: //local.cnspec.io/policies/implicit-props/properties/xtesting
        mql: return true
        for:
        - mrn: //local.cnspec.io/queries/query-1/properties/xtesting

queries:
  - uid: query-1
    title: testing
    mql: props.xtesting == true
    props:
    - mrn: //local.cnspec.io/queries/query-1/properties/xtesting
      type: boolean
```

Fixes #1843

Requires https://github.com/mondoohq/cnquery/pull/6011