🛑 rename require.name => require.provider
After some more feedback in the PR we have decided to rename the "name" field in the new "require" clause to "provider". The current approach to "require" is fully provider-centric, but it will expand in the future. With this approach we avoid polluting the global namespace and avoid adding more layers in the configuration structure. The downside is that we will carry more keys on this layer, which have to be checked for mutual exclusivity. A tradeoff which requires more work on the processor in the future, rather than the user.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>