⭐ policy require providers (#1803)
* ⭐ policy require providers
* 🧹 respect auto-update flag when we check policy provider requirement
* 🧹 use debug for message that the policy is not defining requires properly
* 🧹 add lint check to verify that provider requirements are defined
* 🧹 update cli test
* 🧹 move SortContents into the bundle extension file
* 🧹 fix marshalling of human time and also handle group types and query actions properly
* 🧹 rename lint options for policy linting AutoUpdateProviders to SkipProviderDownload
* 🧹 reuse existing provider installation
* ⭐️ add require field to policies
* 🧹 add comment for explain why we support auto-update disablement for linting
* 🛑 rename require.name => require.provider

After some more feedback in the PR we have decided to rename the "name" field in the new "require" clause to "provider". The current approach to "require" is fully provider-centric, but it will expand in the future. With this approach we avoid polluting the global namespace and avoid adding more layers in the configuration structure. The downside is that we will carry more keys on this layer, which have to be checked for mutual exclusivity. A tradeoff which requires more work on the processor in the future, rather than the user.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>
Co-authored-by: Christoph Hartmann <chris@lollyrock.com>